pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "idwx"
version = "0.1.0"
description = "Idaho weather seasonal frost modeling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "pyarrow>=14",
  "pyyaml>=6.0",
  "typer>=0.12",
  "scikit-learn>=1.4",
  "joblib>=1.3",
]

[project.scripts]
idwx = "idwx.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["idwx"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "ruff>=0.6",
]
